api/requirements.txt
fastapi
uvicorn

httpx>=0.24.1
langchain_core>=0.3.19
langchain_openai>=0.2.9
langchain_deepseek==0.1.2
langsmith>=0.2.11

openai>=1.58.1
pandas==2.2.3

pydantic==2.7.4
python-dotenv

pocketbase
pydub
apscheduler